requirements.txt
flask>=3.0
gensim>=4.3
requests>=2.31
unidecode>=1.3
pywebview>=5.0
rapidfuzz>=3.0
pytest>=8.0
